kafka_streaming_test/requirements.txt
asttokens==3.0.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.12
decorator==5.2.0
exceptiongroup==1.2.2
executing==2.2.0
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.68.0
gspread==6.1.4
httplib2==0.22.0
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kafka-python==2.0.4
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.2
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.50
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
py4j==0.10.9.7
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.19.1
pyparsing==3.2.1
pyspark==3.5.4
python-dateutil==2.9.0.post0
pywin32==308
pyzmq==26.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
